Fully convolutional generator
This is a fully convolutional encoder-decoder type implementation of your generator, as mentioned in my review.

I only changed the `create_generator` function. The only change you need to do elsewhere is to pass the latent variables as a 4d tensor of shape (1,ni//8,nj//8,25) instead of a 1d tensor of shape (100,), where (ni,nj,1) is the dimension of your condition tensor. For example with your training samples the condition has the shape (16,16,1) so the shape of the latent tensor should be (1,2,2,25).

I tested that this produces the correct output on different input shapes but I didn't try training the network so I don't know if it actually works!